+++ lib/mgmg/search.rb
@@ -262,24 +262,25 @@ module Enumerable
 end
 
 module Mgmg
-	module_function def find_lowerbound(a, b, para, start, term, smith_min_a: nil, smith_min_b: nil, min_smith: false)
+	module_function def find_lowerbound(a, b, para, start, term, smith_min_a: nil, smith_min_b: nil, armor_min_a: nil, armor_min_b: nil, min_smith: false)
 		if term <= start
 			raise ArgumentError, "start < term is needed, (start, term) = (#{start}, #{term}) are given"
 		end
 		ira, irb = a.ir, b.ir
-		sca, scb = a.search(para, start, smith_min_a, min_smith: min_smith, irep: ira), b.search(para, start, smith_min_b, min_smith: min_smith, irep: irb)
+		sca, scb = a.search(para, start, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira), b.search(para, start, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 		ea, eb = Mgmg.exp(*sca), Mgmg.exp(*scb)
 		if eb < ea || ( ea == eb && ira.para_call(para, *sca) < irb.para_call(para, *scb) )
-			a, b, ira, irb, sca, scb, ea, eb, smith_min_a, smith_min_b = b, a, irb, ira, scb, sca, eb, ea, smith_min_b, smith_min_a
+			a, b, ira, irb, sca, scb, ea, eb = b, a, irb, ira, scb, sca, eb, ea
+			smith_min_a, smith_min_b, armor_min_a, armor_min_b = smith_min_b, smith_min_a, armor_min_b, armor_min_a
 		end
 		tag = ira.para_call(para, *sca) + 1
-		sca, scb = a.search(para, term, smith_min_a, min_smith: min_smith, irep: ira), b.search(para, term, smith_min_b, min_smith: min_smith, irep: irb)
+		sca, scb = a.search(para, term, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira), b.search(para, term, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 		ea, eb = Mgmg.exp(*sca), Mgmg.exp(*scb)
 		if ea < eb || ( ea == eb && irb.para_call(para, *scb) < ira.para_call(para, *sca) )
 			raise Mgmg::SearchCutException
 		end
 		while tag < term
-			sca, scb = a.search(para, tag, smith_min_a, min_smith: min_smith, irep: ira), b.search(para, tag, smith_min_b, min_smith: min_smith, irep: irb)
+			sca, scb = a.search(para, tag, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira), b.search(para, tag, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 			ea, eb = Mgmg.exp(*sca), Mgmg.exp(*scb)
 			pa, pb = ira.para_call(para, *sca), irb.para_call(para, *scb)
 			if eb < ea
@@ -297,52 +298,53 @@ module Mgmg
 		raise UnexpectedError
 	end
 	
-	module_function def find_upperbound(a, b, para, start, term, smith_min_a: nil, smith_min_b: nil, min_smith: false)
+	module_function def find_upperbound(a, b, para, start, term, smith_min_a: nil, smith_min_b: nil, armor_min_a: nil, armor_min_b: nil, min_smith: false)
 		if start <= term
 			raise ArgumentError, "term < start is needed, (start, term) = (#{start}, #{term}) are given"
 		end
 		ira, irb = a.ir, b.ir
-		sca, scb = a.search(para, start, smith_min_a, min_smith: min_smith, irep: ira), b.search(para, start, smith_min_b, min_smith: min_smith, irep: irb)
+		sca, scb = a.search(para, start, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira), b.search(para, start, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 		ea, eb = Mgmg.exp(*sca), Mgmg.exp(*scb)
 		if ea < eb || ( ea == eb && irb.para_call(para, *scb) < ira.para_call(para, *sca) )
-			a, b, ira, irb, sca, scb, ea, eb, smith_min_a, smith_min_b = b, a, irb, ira, scb, sca, eb, ea, smith_min_b, smith_min_a
+			a, b, ira, irb, sca, scb, ea, eb = b, a, irb, ira, scb, sca, eb, ea
+			smith_min_a, smith_min_b, armor_min_a, armor_min_b = smith_min_b, smith_min_a, armor_min_b, armor_min_a
 		end
 		tagu = ira.para_call(para, *sca)
 		sca[-1] -= 2
 		tagl = ira.para_call(para, *sca)
-		sca, scb = a.search(para, term, smith_min_a, min_smith: min_smith, irep: ira), b.search(para, term, smith_min_b, min_smith: min_smith, irep: irb)
+		sca, scb = a.search(para, term, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira), b.search(para, term, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 		ea, eb = Mgmg.exp(*sca), Mgmg.exp(*scb)
 		if eb < ea || ( ea == eb && ira.para_call(para, *sca) < irb.para_call(para, *scb) )
 			raise Mgmg::SearchCutException
 		end
 		while term < tagu
 			ret = nil
-			sca = a.search(para, tagl, smith_min_a, min_smith: min_smith, irep: ira)
+			sca = a.search(para, tagl, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira)
 			next_tagu, next_sca = tagl, sca
-			scb = b.search(para, tagl, smith_min_b, min_smith: min_smith, irep: irb)
+			scb = b.search(para, tagl, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 			while tagl < tagu
 				ea, eb = Mgmg.exp(*sca), Mgmg.exp(*scb)
 				pa, pb = ira.para_call(para, *sca), irb.para_call(para, *scb)
 				if ea < eb
 					ret = tagl
-					sca = a.search(para, pa + 1, smith_min_a, min_smith: min_smith, irep: ira)
+					sca = a.search(para, pa + 1, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira)
 					tagl = ira.para_call(para, *sca)
-					scb = b.search(para, tagl, smith_min_b, min_smith: min_smith, irep: irb)
+					scb = b.search(para, tagl, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 				elsif ea == eb
 					if pb < pa
 						ret = tagl
-						sca = a.search(para, pa + 1, smith_min_a, min_smith: min_smith, irep: ira)
+						sca = a.search(para, pa + 1, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira)
 						tagl = ira.para_call(para, *sca)
-						scb = b.search(para, tagl, smith_min_b, min_smith: min_smith, irep: irb)
+						scb = b.search(para, tagl, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 					else
-						scb = b.search(para, pb + 1, smith_min_b, min_smith: min_smith, irep: irb)
+						scb = b.search(para, pb + 1, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 						tagl = irb.para_call(para, *scb)
-						sca = a.search(para, tagl, smith_min_a, min_smith: min_smith, irep: ira)
+						sca = a.search(para, tagl, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira)
 					end
 				else
-					sca = a.search(para, pa + 1, smith_min_a, min_smith: min_smith, irep: ira)
+					sca = a.search(para, pa + 1, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira)
 					tagl = ira.para_call(para, *sca)
-					scb = b.search(para, tagl, smith_min_b, min_smith: min_smith, irep: irb)
+					scb = b.search(para, tagl, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb)
 				end
 			end
 			if ret.nil?
@@ -353,8 +355,8 @@ module Mgmg
 					tagl = term
 				end
 			else
-				pa = ira.para_call(para, *a.search(para, ret+1, smith_min_a, min_smith: min_smith, irep: ira))
-				pb = irb.para_call(para, *b.search(para, ret+1, smith_min_b, min_smith: min_smith, irep: irb))
+				pa = ira.para_call(para, *a.search(para, ret+1, smith_min_a, armor_min_a, min_smith: min_smith, irep: ira))
+				pb = irb.para_call(para, *b.search(para, ret+1, smith_min_b, armor_min_b, min_smith: min_smith, irep: irb))
 				return [ret, [pa, pb].min]
 			end
 		end